Work in progress on replayer. First program replayed without errors!
git-svn-id: http://pleiad.dcc.uchile.cl/svn/tod/trunk@2443 52faa79f-a2d0-0310-bd40-81a52920883b

+++ core/src/debugger/tod/impl/replay/LongDeltaReceiver.java
@@ -45,7 +45,7 @@ public class LongDeltaReceiver
 	
 	public long receiveFull(_ByteBuffer aBuffer)
 	{
-		itsValue = aBuffer.getInt();
+		itsValue = aBuffer.getLong();
 		return itsValue;
 	}
